Plugins: Fixed compilation warnings
Former-commit-id: b8c298ad993c49d40148bbc176a619f109c29252

+++ plugins/processing/data-generation/src/box-algorithms/ovpCBoxAlgorithmNoiseGenerator.cpp
@@ -73,13 +73,13 @@ boolean CNoiseGenerator::initialize(void)
 	IMatrix &l_rSampleMatrix = *ip_pSignalMatrix;
 
 	l_rSampleMatrix.setDimensionCount(2);
-	l_rSampleMatrix.setDimensionSize(0,(uint32)m_ui64ChannelCount);
-	l_rSampleMatrix.setDimensionSize(1,(uint32)m_ui64GeneratedEpochSampleCount);
-	for(uint64 i=0;i<m_ui64ChannelCount;i++)
+	l_rSampleMatrix.setDimensionSize(0,static_cast<uint32>(m_ui64ChannelCount));
+	l_rSampleMatrix.setDimensionSize(1,static_cast<uint32>(m_ui64GeneratedEpochSampleCount));
+	for(uint32 i=0;i<static_cast<uint32>(m_ui64ChannelCount);i++)
 	{
 		char l_sBuffer[64];
 		sprintf(l_sBuffer, "Noise %d", i);
-		l_rSampleMatrix.setDimensionLabel(0, static_cast<uint32>(i), l_sBuffer);
+		l_rSampleMatrix.setDimensionLabel(0, i, l_sBuffer);
 	}
 
 	return true;
